refactor: Remove ChatInstance component and integrate its logic directly into Chat component
